Changed casting of variable.  CAS-11027
git-svn-id: https://svn.cv.nrao.edu/svn/casa/branches/project/pipeline@42523 ac9c0da5-131f-0410-9009-a9c195f183e6

+++ pipeline/h/tasks/importdata/fluxes.py
@@ -431,7 +431,7 @@ def import_flux(output_dir, observing_run, filename=None):
             age = match.group('age') if match else None
             if age:
                 try:
-                    age = int(age)
+                    age = float(age)
                 except ValueError:
                     age = None
